SIL.Llvm

Documentation

haskFun :: FunPtr (IO (Ptr Int64)) -> IO (Ptr Int64)

data RunResult

Constructors

RunResult

Fields

resultValue :: Int64

resultNumPairs :: Int64

run :: JITConfig -> WordPtr -> IO RunResult

convertPairs :: RunResult -> IO NExpr

Recursively follow all integers by interpreting them as indices in the pair array until a 0 is found.

makeModule :: NExpr -> Module

data DebugModule

Constructors

DebugModule Module

resolver :: CompileLayer l => l -> SymbolResolver

withTargetMachine :: (TargetMachine -> IO a) -> IO a

debugLog :: JITConfig -> String -> IO ()

optimizeModule :: JITConfig -> Module -> IO ()

data JITConfig

Constructors

JITConfig

Fields

debugOutput :: !Bool

timingOutput :: !Bool

optimizerLevel :: !OptimizerLevel

defaultJITConfig :: JITConfig

data OptimizerLevel

Constructors

None

One

Two

Three

optimizerLevelToWord :: OptimizerLevel -> Word

evalJIT :: JITConfig -> Module -> IO (Either String RunResult)

printTimings :: TimeSpec -> TimeSpec -> TimeSpec -> TimeSpec -> TimeSpec -> TimeSpec -> TimeSpec -> IO ()

fmtTS :: TimeSpec -> String

intT :: Type

boolT :: Type

intPtrT :: Type

funT :: Type

voidFunT :: Type

setjmpT :: Type

longjmpT :: Type

gcMallocT :: Type

funPtrT :: Type

pairT :: Type

pairPtrT :: Type

heapIndexN :: Name

heapIndex :: Definition

resultStructureN :: Name

resultStructureT :: Type

resultStructure :: Definition

zero :: Operand

one :: Operand

zero32 :: Operand

one32 :: Operand

two32 :: Operand

type BuilderInternal = State (Map ByteString Operand, Int)

type SILBuilder = IRBuilderT (ModuleBuilderT BuilderInternal)

doFunction :: NExpr -> SILBuilder Operand

Wrap the argument in a function that accepts an integer (index to environment) as its argument.

doAnonFunction :: SILBuilder Operand -> SILBuilder Operand

pairOffC :: Operand

heapIndexC :: Operand

gcMallocPair :: (MonadModuleBuilder m, MonadIRBuilder m) => m Operand

makePair :: (MonadModuleBuilder m, MonadIRBuilder m) => Operand -> Operand -> m Operand

makePair a b allocates a new pair (a,b) at the current heap index and increments the heap index.

goLeft :: (MonadModuleBuilder m, MonadFix m) => m Operand

goLeft p loads the first element of the pair p.

goRight :: (MonadModuleBuilder m, MonadFix m) => m Operand

goRight p loads the second element of the pair p.

doLeft :: MonadIRBuilder m => Operand -> m Operand

doRight :: MonadIRBuilder m => Operand -> m Operand

envC :: Operand

lComment :: a -> (a, MDRef MDNode)

toLLVM :: Map FragIndex ExprFrag -> ExprFrag -> SILBuilder Operand

resultC :: Operand